feat: implement VTP flag.
logion-network/logion-internal#678
