+++ test/unit/model/locrequest.model.spec.ts
@@ -342,6 +342,25 @@ describe("LocRequestAggregateRoot", () => {
         givenRequestWithStatus('REQUESTED');
         expect(() => whenSubmitting()).toThrowError();
     });
+
+    it("sets verified third party flag if closed Identity LOC", () => {
+        givenRequestWithStatus('CLOSED');
+        request.locType = "Identity";
+        request.setVerifiedThirdParty(true);
+        expect(request.verifiedThirdParty).toBe(true);
+    });
+
+    it("fails setting verified third party flag if non-closed", () => {
+        givenRequestWithStatus('OPEN');
+        request.locType = "Identity";
+        expect(() => request.setVerifiedThirdParty(true)).toThrowError();
+    });
+
+    it("fails setting verified third party flag if non-Identity", () => {
+        givenRequestWithStatus('CLOSED');
+        request.locType = "Transaction";
+        expect(() => request.setVerifiedThirdParty(true)).toThrowError();
+    });
 });
 
 describe("LocRequestAggregateRoot (metadata)", () => {
